Split stale review diagnostics assembly
