+++ src/supervisor/supervisor-detailed-status-assembly.ts
@@ -6,10 +6,10 @@ import {
   buildCodexConnectorDiagnosticBundle,
 } from "./supervisor-status-review-bot";
 import {
-  buildStaleReviewBotThreadDiagnostics,
   buildStaleReviewBotRemediation,
   currentHeadVerifiedRepairResidueArtifactEvidenceSummary,
 } from "./stale-review-bot-remediation";
+import { buildStaleReviewBotThreadDiagnostics } from "./stale-review-bot-diagnostics";
 import {
   formatStaleReviewBotRemediationLine,
   formatStaleReviewBotRepairTargetLine,
